+++ main.cpp
@@ -17,11 +17,11 @@ double figsize_scale = 0.2;
 
 
 double learning_rate = 0.1;
-static int GRID_SIZE = 4;
-static int RESOLUTION = 5;
-static int LOWER_RES = 2;
-static int UPPER_RES = 3;
-static int BATCH_SIZE = 1000;
+static size_t GRID_SIZE = 4;
+static size_t RESOLUTION = 5;
+static size_t LOWER_RES = 2;
+static size_t UPPER_RES = 3;
+static size_t BATCH_SIZE = 1000;
 
 /*
  * Reads the MNIST dataset from binary file located at "./data/train-images-idx3-ubyte"
@@ -83,7 +83,7 @@ auto data = get_data();
 template <typename T>
 CubeArray<T> get_batch(size_t batch_size){
     std::vector<std::vector<size_t>> batch_indices(batch_size, std::vector<size_t>(3));
-    for(int i = 0; i < batch_size; ++i) {
+    for(size_t i = 0; i < batch_size; ++i) {
         std::vector<size_t> temp;
         batch_indices[i][0] = ((int)((get_rand() * 60000.)));
         // todo hardcoded shapes
@@ -93,7 +93,7 @@ CubeArray<T> get_batch(size_t batch_size){
 
     std::vector<std::vector<std::vector<T>>> batch;
 
-    for (int i = 0; i < batch_indices.size(); ++i) {
+    for (size_t i = 0; i < batch_indices.size(); ++i) {
         auto dt = data[batch_indices[i][0]];
         batch.emplace_back( dt.get_slices(batch_indices[i][1] - LOWER_RES, batch_indices[i][1] + UPPER_RES, batch_indices[i][2] - LOWER_RES, batch_indices[i][2] + UPPER_RES));
     }
@@ -113,27 +113,27 @@ void experiment(const char subfigure, double sigma, double lambda_, size_t nbatc
 
     Model<T> model(sigma, lambda_, GRID_SIZE, RESOLUTION, learning_rate);
 
-    short num_threads = 10;
+    size_t num_threads = 10;
     std::vector<std::thread> threads(num_threads);
     for (size_t i = 0; i < nbatches; i++){
         auto start = std::chrono::high_resolution_clock::now();
         CubeArray<T> batch = get_batch<double>(BATCH_SIZE);
         std::vector<CubeArray<T>> cubs;
         cubs.reserve(num_threads);
-        for (int j = 0; j < num_threads; ++j) {
+        for (size_t j = 0; j < num_threads; ++j) {
             cubs.push_back(model.w);
             threads[j] = std::thread([j, &batch, &model, &cubs, num_threads]{
                 auto m = model;
-                for (int k = (BATCH_SIZE / num_threads) * j; k < (BATCH_SIZE / num_threads) * (j + 1); ++k) {
+                for (size_t k = (BATCH_SIZE / num_threads) * j; k < (BATCH_SIZE / num_threads) * (j + 1); ++k) {
                     m.update(batch[k]);
                 }
                 cubs[j] = (m.w);
             });
         }
-        for (int j = 0; j < num_threads; ++j) {
+        for (size_t j = 0; j < num_threads; ++j) {
             threads[j].join();
         }
-        for (int j = 1; j < num_threads; ++j) {
+        for (size_t j = 1; j < num_threads; ++j) {
             cubs[0] += cubs[j];
         }
         model.w = cubs[0] / num_threads;
